Fixed a bug in the B+Tree that got introduced when I started using cached indexes. Cleaned up the implementation of the Memory Stream. Added more debug code to catch errors.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS11]$/openHistorian/Main;C23026

+++ Source/Libraries/Tests/openHistorian.UnitTests/Engine/ArchiveWriters/StageWriterTest.cs
@@ -43,7 +43,7 @@ namespace openHistorian.Engine.ArchiveWriters
 
         void FinalizeArchiveFile(RolloverArgs args)
         {
-            m_pointsRead = args.File.Count();
+            m_pointsRead = (int)args.File.Count();
             m_sequenceNumber = args.SequenceNumber;
             Assert.NotNull(args.File);
         }
